requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==4.2.19
django-crispy-forms==2.3
django-filebrowser==4.0.3
django-grappelli==3.0.9
django-stubs==5.1.3
django-stubs-ext==5.1.3
django-tinymce==4.1.0
pillow==11.1.0
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
